Extract common parts to AbstractWindowsConsoleWriter
